First version of path planner
The first version of the path planner is able
to stay within its own lane. It is able to stay
within the speed limit. It is able to slow down
when too close to another car and it is able to
speed up when there isn't a car ahead of it. It
can merge into the left lane, but cannot merge
into the right lane and doesn't check if it is
safe to merge into the left lane. Code was
referenced from udacity video walkthrough. This
code should help anyone who is trying to get
started with the project.