Added Path Planner Code to project
My Path Planner implementation is based on the code
from Udacity's walkthrough video. My Path Planner
goes through 3 steps to navigate the car safely around
the virtual highway: prediction, behavior planning and
trajectory generation. In the prediction step, the
planner must identify which lanes (left, right, same)
are being occupied by other cars using sensor fusion
data. For instance, there is a vehicle in the same lane
ahead of our car and our car is too close to that other
vehicle. In the behavior planning step, the planner
must make decisions when our car is too close to another
car, such as merging into the appropriate lane or slowing
down since there is not an available lane to merge into.
In the trajectory generation step, it creates smooth,
drivable and collision-free trajectories for the motion
controller to follow. The spline library is used to
create these kind of trajectories.