Added project overview and how to run demo
I added a project overview, path planning background,
contents of the project, reflection on how to generate
paths, how to run the demo and test the path planner
to see it navigate a car on a virtual highway safely.
I also added resources that I found helpful for
finishing the project, such as Udacity SDCE and github
repos.